added detection of equivalent edge pairs; savings of 90% or more in BEM matrix assembly for structured meshes

+++ libs/libSubstrate/SommerfeldIntegrator.cc
@@ -325,7 +325,7 @@ void SommerfeldIntegrate(integrand f, void *fdata, unsigned zfdim,
       break;
    };
 
-  for(int nf=0; nf<zfdim; nf++)
+  for(size_t nf=0; nf<zfdim; nf++)
    { Integral[nf] += SValues[nf];
      Error[nf]    += abs(SValues[nf] - LastSValues[nf]);
    };
